Gussy: Fixed uint16 on LEDs
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@74 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ flight/PiOS/STM32F10x/pios_led.c
@@ -35,7 +35,7 @@
 
 /* Local Variables */
 static GPIO_TypeDef* LED_GPIO_PORT[NUM_LED] = {LED1_GPIO_PORT, LED2_GPIO_PORT};
-static const uint16_t LED_GPIO_PIN[NUM_LED] = {LED1_GPIO_PIN, LED2_GPIO_PIN};
+static const uint32_t LED_GPIO_PIN[NUM_LED] = {LED1_GPIO_PIN, LED2_GPIO_PIN};
 static const uint32_t LED_GPIO_CLK[NUM_LED] = {LED1_GPIO_CLK, LED2_GPIO_CLK};
